requirements.txt
apache-airflow==2.6.1
Scrapy==2.9.0
s3fs==2023.5.0
botocore==1.29.140
apache-airflow[amazon]
TextBlob
Tweepy